Move buildNextTasks() to taskProcessor for better separation of concerns.
