+++ lib/queue.js
@@ -11,7 +11,6 @@ const request = require('./request');
 const taskProcessor = require('./taskProcessor');
 const taskHelper = require('./helpers/task');
 const log = require("./helpers/logger");
-const operator = require("./helpers/operator");
 
 const maxConcurrent = process.env.MAX_PARALLEL_TASKS ||
     (config.has('app.max_parallel_tasks') ? config.get('app.max_parallel_tasks') : 3);
@@ -33,7 +32,7 @@ const responseHandler = function(task, response){
 
     if (_.has(task, 'then')) {
 
-        let nextTasks = buildNextTasks(task, response);
+        let nextTasks = taskProcessor.buildNextTasks(task, response);
 
         //Add new tasks to queue. Return a promise that will resolve
         //only once all the items in array have resolved.
@@ -81,92 +80,6 @@ const responseHandler = function(task, response){
 
 
 /**
- * Generate an array of request objects from a response result keys.
- *
- * @param task
- * @param response
- * @returns [{},{}] - array of request objects
- */
-const buildNextTasks = function(task, response){
-
-    let nextTasks = [];
-
-    log.trace("Building next task from THEN object:", JSON.stringify(task.then, undefined, 4));
-
-    response.analysis.results.map(result => {
-
-        let nextTask = _.clone(task.then);
-        let analysisType = response.analysis.analysis_type;
-
-        //FREQDIST responses - build next FILTER from result key
-        if(analysisType === 'freqDist') {
-            log.debug("Building FILTER for freqDist");
-
-            //append cascading filter
-            nextTask.filter = "(" + response.analysis.parameters.target + " "
-                + operator.get(response.analysis.parameters.target) + " \"" + result.key + "\")";
-
-            //parent task had a filter, append
-            if (_.has(task.json, "filter")) {
-                nextTask.filter += " AND (" + task.json.filter + ")";
-            }
-
-            //child task had a filter, append
-            if (_.has(task.then, "filter")) {
-                nextTask.filter += " AND (" + task.then.filter + ")";
-            }
-            
-            //set task id from response result key and inherit
-            //any id value set in the config
-            nextTask.id = _.has(task, 'cache') ? task.cache.mergeKey + "__" + result.key : result.key;
-        }
-
-        // TIMESERIES responses - build next filter START, END from result key
-        if(analysisType === 'timeSeries'){
-            log.debug("Building START & END for timeSeries");
-
-            let endOverride = taskHelper.getEndTs(result.key, response.analysis.parameters.interval);
-
-            //parent task had a filter, append
-            if (_.has(task.json, "filter"))
-                nextTask.filter = "(" + task.json.filter + ")";
-
-            //child task had a filter, append
-            if (_.has(task.then, "filter")) {
-                nextTask.filter += " AND (" + task.then.filter + ")";
-            }
-
-            //only override the start if its less than 32 days as timeSeries can
-            //return an interval start outside of the max period. If that happens
-            //the default obj value will be used ie now - 32 days
-            if(result.key > moment.utc().subtract(32, 'days').unix()){
-                nextTask.start = result.key;
-            }
-            nextTask.end = endOverride;
-
-            log.trace("Generated start: " + nextTask.start + " "
-                + moment.unix(nextTask.start).utc().format("DD-MMM h:mm:ss a"));
-
-            log.trace("Generated end:   " + nextTask.end + " "
-                + moment.unix(nextTask.end).utc().format("DD-MMM h:mm:ss a"));
-
-            //set task id from response result key and inherit
-            //any id value set in the config
-            nextTask.id = _.has(task, 'cache') ? task.cache.mergeKey + "__"
-                + taskHelper.unixToHuman(result.key) : taskHelper.unixToHuman(result.key);
-        }
-
-        if (_.has(task, 'then') && _.has(task, 'cache'))
-            nextTask.parentCache = _.clone(task.cache.cacheId);
-
-        nextTasks.push(nextTask);
-    });
-
-    return taskProcessor.buildRequest(nextTasks, "freqDist");
-};
-
-
-/**
  * queueTask - add task to queue
  *
  * @param task
@@ -192,11 +105,13 @@ const queueTask = function queueTask(task) {
     }).then(response => {
         return responseHandler(task, response);
     }).catch(function (err) {
+
         if (err.error && err.error.error) {
             log.error(JSON.stringify(err.error.error, undefined, 4));
         } else {
-            log.error(JSON.stringify(err, undefined, 4));
+            log.error(err);
         }
+
         process.exit(1);
     });
 };
