+++ lib/taskProcessor.js
@@ -6,6 +6,7 @@ const shortid = require('shortid');
 const taskHelper = require('./helpers/task');
 const cacheHelper = require('./helpers/cache');
 const log = require("./helpers/logger");
+const operator = require("./helpers/operator");
 
 
 /**
@@ -130,6 +131,95 @@ const buildRequest = function buildRequest(tasks, type, mergedName, globalFilter
 };
 
 
+
+/**
+ * buildNextTasks - Generate an array of request
+ * objects from a response result keys.
+ *
+ * @param task
+ * @param response
+ * @returns [{},{}] - array of request objects
+ */
+const buildNextTasks = function(task, response){
+
+    let nextTasks = [];
+
+    log.trace("Building next task from THEN object:", JSON.stringify(task.then, undefined, 4));
+
+    response.analysis.results.map(result => {
+
+        let nextTask = _.clone(task.then);
+        let analysisType = response.analysis.analysis_type;
+
+        //FREQDIST responses - build next FILTER from result key
+        if(analysisType === 'freqDist') {
+            log.debug("Building FILTER for freqDist");
+
+            //append cascading filter
+            nextTask.filter = "(" + response.analysis.parameters.target + " "
+                + operator.get(response.analysis.parameters.target) + " \"" + result.key + "\")";
+
+            //parent task had a filter, append
+            if (_.has(task.json, "filter")) {
+                nextTask.filter += " AND (" + task.json.filter + ")";
+            }
+
+            //child task had a filter, append
+            if (_.has(task.then, "filter")) {
+                nextTask.filter += " AND (" + task.then.filter + ")";
+            }
+
+            //set task id from response result key and inherit
+            //any id value set in the config
+            nextTask.id = _.has(task, 'cache') ? task.cache.mergeKey + "__" + result.key : result.key;
+        }
+
+        // TIMESERIES responses - build next filter START, END from result key
+        if(analysisType === 'timeSeries'){
+            log.debug("Building START & END for timeSeries");
+
+            let endOverride = taskHelper.getEndTs(result.key, response.analysis.parameters.interval);
+
+            //parent task had a filter, append
+            if (_.has(task.json, "filter"))
+                nextTask.filter = "(" + task.json.filter + ")";
+
+            //child task had a filter, append
+            if (_.has(task.then, "filter")) {
+                nextTask.filter += " AND (" + task.then.filter + ")";
+            }
+
+            //only override the start if its less than 32 days as timeSeries can
+            //return an interval start outside of the max period. If that happens
+            //the default obj value will be used ie now - 32 days
+            if(result.key > moment.utc().subtract(32, 'days').unix()){
+                nextTask.start = result.key;
+            }
+            nextTask.end = endOverride;
+
+            log.trace("Generated start: " + nextTask.start + " "
+                + moment.unix(nextTask.start).utc().format("DD-MMM h:mm:ss a"));
+
+            log.trace("Generated end:   " + nextTask.end + " "
+                + moment.unix(nextTask.end).utc().format("DD-MMM h:mm:ss a"));
+
+            //set task id from response result key and inherit
+            //any id value set in the config
+            nextTask.id = _.has(task, 'cache') ? task.cache.mergeKey + "__"
+            + taskHelper.unixToHuman(result.key) : taskHelper.unixToHuman(result.key);
+        }
+
+        if (_.has(task, 'then') && _.has(task, 'cache'))
+            nextTask.parentCache = _.clone(task.cache.cacheId);
+
+        nextTasks.push(nextTask);
+    });
+
+    return buildRequest(nextTasks, "freqDist");
+};
+
+
+
 /**
  * processNativeNested - correctly format any native
  * nested config that has been applied to the
@@ -165,6 +255,7 @@ const processNativeNested = function processNativeNested(taskObj, reqObj){
     }
     return reqObj;
 };
+
 
 
 /**
@@ -230,4 +321,5 @@ const loadConfigTasks = function loadConfigTasks(configTest) {
 };
 
 exports.loadConfigTasks = loadConfigTasks;
+exports.buildNextTasks = buildNextTasks;
 exports.buildRequest = buildRequest;
